+++ Mocks/True DGTL/DGTLBakcendMock/DGTLBakcendMock.Common/DTO/SecurityList/DailySettlementPrice.cs
@@ -17,6 +17,6 @@ namespace DGTLBackendMock.Common.DTO.SecurityList
 
         public string DSPTime { get; set; }
 
-        public decimal? Price { get; set; }
+        public double? Price { get; set; }
     }
 }
